requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-barcode
python-slugify
reportlab
Pillow
fastapi-mail
psycopg2-binary
fpdf2
cloudinary
requests